move NodeView to reportviews
