+++ toponetx/classes/simplicial_complex.py
@@ -16,8 +16,7 @@ from networkx import Graph
 from scipy.sparse import coo_matrix, dok_matrix
 
 from toponetx.classes.complex import Complex
-from toponetx.classes.node import NodeView
-from toponetx.classes.reportviews import SimplexView
+from toponetx.classes.reportviews import SimplexView, NodeView
 from toponetx.classes.simplex import Simplex
 from toponetx.exception import TopoNetXError
